adding blog without a proper token ei toimi vielä...

+++ osa4/controllers/blogs.js
@@ -2,6 +2,8 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
+
+
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', {username:1, name:1})
         response.json(blogs)
@@ -9,11 +11,13 @@ blogsRouter.get('/', async (request, response) => {
 
   blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
-    const user = await User.findById(body.userId)
-    // console.log(user)
+
+
+    //console.log(body)
+    const user = request.user
+    console.log(user)
 
     let likes = 0
-
     if(body.likes) {
       likes = body.likes
     }
@@ -30,10 +34,10 @@ blogsRouter.get('/', async (request, response) => {
 
     try{
     const savedBlog = await blog.save()
-    console.log(savedBlog.id)
+    //console.log(savedBlog.id)
     user.blogs = user.blogs.concat(savedBlog.id)
     await user.save()
-    console.log(user.blogs)
+    //console.log(user.blogs)
 
     response.status(201).json(savedBlog)
 
@@ -46,9 +50,17 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+    const body = request.body
+  
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
   try{
+    if ( blog.user.toString() === user.id.toString() ) {
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'Cannot delete other users blogs' })
+    }
   } catch(exception) {
     next(exception)
   }
